requirements-huggingface.txt
torch>=2.0.0
pillow>=9.4.0
streamlit>=1.24.0
opencv-python-headless>=4.7.0
pydicom>=2.3.1
numpy>=1.24.3
pandas>=2.0.2
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
gradio>=3.34.0
huggingface_hub>=0.15.1 